\name{MYHfusionRcpp-package}
\alias{MYHfusionRcpp-package}
\alias{MYHfusionRcpp}
\docType{package}
\title{
\packageTitle{MYHfusionRcpp}
}
\description{
\packageDescription{MYHfusionRcpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MYHfusionRcpp}
\packageIndices{MYHfusionRcpp}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{MYHfusionRcpp}

Maintainer: \packageMaintainer{MYHfusionRcpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
