\name{MYHfusionRcpp-package}
\alias{MYHfusionRcpp-package}
\alias{MYHfusionRcpp}
\docType{package}
\title{
\packageTitle{MYHfusionRcpp}
}
\description{
\packageDescription{MYHfusionRcpp}
}
\details{
\packageDESCRIPTION{MYHfusionRcpp}
\packageIndices{MYHfusionRcpp}

This package provides functions for
kernel smoothing,
MLE for generalized linear models,
panel data analysis,
survival analysis,
using Rcpp for high performance.
}
\author{
\packageAuthor{MYHfusionRcpp}

Maintainer: \packageMaintainer{MYHfusionRcpp}
}
\keyword{ package }
\examples{
## Simple example
# library(MYHfusionRcpp)
# ... (add your example code here if you have one)
}
